Move Mapper to own package
