+++ cameraview/src/androidTest/java/com/otaliastudios/cameraview/engine/MapperTest.java
@@ -9,6 +9,7 @@ import com.otaliastudios.cameraview.controls.Facing;
 import com.otaliastudios.cameraview.controls.Flash;
 import com.otaliastudios.cameraview.controls.Hdr;
 import com.otaliastudios.cameraview.controls.WhiteBalance;
+import com.otaliastudios.cameraview.engine.mappers.Mapper;
 
 import androidx.test.ext.junit.runners.AndroidJUnit4;
 import androidx.test.filters.SmallTest;
@@ -21,7 +22,7 @@ import static org.junit.Assert.*;
 
 @RunWith(AndroidJUnit4.class)
 @SmallTest
-public class Mapper1Test extends BaseTest {
+public class MapperTest extends BaseTest {
 
     private Mapper mapper = Mapper.get(Engine.CAMERA1);
